% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Viewer_module.R
\name{data_Viewer}
\alias{data_Viewer}
\title{The corresponding server function to data_ViewerUI().}
\usage{
data_Viewer(input, output, session, volumes = volumes)
}
\description{
The corresponding server function to data_ViewerUI().
}
\author{
Daniel Braas
}
