% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Viewer_module.R
\name{data_ViewerUI}
\alias{data_ViewerUI}
\title{This UI function contains allows both viewing as well as production of graphs.
The finished UI will contain multiple possibilities starting with bar
graphs of relative or absolute data, as well as}
\usage{
data_ViewerUI(id, label, Title)
}
\arguments{
\item{id}{The module ID.}

\item{label}{The label to be displayed depending on the type of data.}

\item{Title}{The text field to be displayed when opening the file.}
}
\value{
A bar graph depicting metabolomics data based on the selected pathway.
}
\description{
This UI function contains allows both viewing as well as production of graphs.
The finished UI will contain multiple possibilities starting with bar
graphs of relative or absolute data, as well as
}
\author{
Daniel Braas
}
