% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extracts.module.R
\name{extract_analysis}
\alias{extract_analysis}
\title{This is the corresponding server function to the extract_analysisUI function.}
\usage{
extract_analysis(input, output, session, volumes = volumes)
}
\value{
The processed data in the selected folder.
}
\description{
This is the corresponding server function to the extract_analysisUI function.
}
\author{
Daniel Braas
}
