% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extracts.module.R
\name{extract_analysisUI}
\alias{extract_analysisUI}
\title{This function produces the UI to process data analyzed with TraceFinder.}
\usage{
extract_analysisUI(id)
}
\arguments{
\item{id}{The module id used in the app.}
}
\value{
The processed data in the selected folder.
}
\description{
This function produces the UI to process data analyzed with TraceFinder.
}
\author{
Daniel Braas
}
