% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{make_labeled}
\alias{make_labeled}
\title{This function produces a data frame with percent labeled data.}
\usage{
make_labeled(DF, Title = "", folder = getwd())
}
\arguments{
\item{DF}{The input data. This should be a data frame that contains MID data generated
with the 'make_MID' function.}

\item{Title}{The title to be used.}

\item{folder}{The folder where the data is supposed to be saved. Default is the current directory.}
}
\value{
A data frame with percent labeled data.
}
\description{
This function produces a data frame with percent labeled data.
}
\author{
Daniel Braas
}
