% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{make_MID}
\alias{make_MID}
\title{This function produces a data frame with mass isotopologue distribution (MID) data that is corrected for naturally occurring 13C.}
\usage{
make_MID(DF, Title = "", folder = getwd())
}
\arguments{
\item{DF}{The input data. This should be a data frame with Name, Iso, Condition, Exp,
Value, Used_ID, KEGG.ID, Nr.C, Rt and Formula columns.}

\item{Title}{The title to be used.}

\item{folder}{The folder where the data is supposed to be saved. Default is thecurrent directory.}
}
\value{
A data frame with MID data.
}
\description{
This function produces a data frame with mass isotopologue distribution (MID) data that is corrected for naturally occurring 13C.
}
\author{
Daniel Braas
}
