% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_graphs.R
\name{make_PCA}
\alias{make_PCA}
\title{This function takes in a data matrix and produces a PCA plot and a correlation circle plot.}
\usage{
make_PCA(
  matrix,
  info,
  a = 1,
  b = 2,
  cutoff = 0.5,
  Title = "",
  folder = getwd()
)
}
\arguments{
\item{matrix}{The data matrix to be used.}

\item{info}{A data frame with information about the Conditions and Cell.Number}

\item{a}{The PC used for the x-axis.}

\item{b}{The PC used for the y-axis.}

\item{cutoff}{The cutoff for correlation.}

\item{Title}{The title to be used. This title will also be part of the file name.}

\item{folder}{The folder where the data is supposed to be saved. Default is the current directory.}
}
\value{
A pdf file with a scree plot, a pair plot showing the first five PCs as well as a PCA plot with PCs a and b as specified in the function call and the corresponding top 30 loadings as bar plot.
}
\description{
This function takes in a data matrix and produces a PCA plot and a correlation circle plot.
}
\author{
Daniel Braas
}
