% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_graphs.R
\name{make_pheatmap}
\alias{make_pheatmap}
\title{This function takes in a data matrix and produces an annotated heatmap using the pheatmap package.}
\usage{
make_pheatmap(
  matrix,
  samples = info,
  heat.color = c("navy", "grey90", "firebrick3"),
  Norv = 1,
  Title,
  folder = getwd()
)
}
\arguments{
\item{matrix}{The data matrix to be used}

\item{samples}{A data frame with information about the Conditions and Cell.Number}

\item{heat.color}{The color scheme to be used for the heatmap.}

\item{Norv}{The internal standard}

\item{Title}{The title to be used. This title will also be part of the file name.}

\item{folder}{The folder where the data is supposed to be saved. Default is the current directory.}
}
\value{
An annotated heatmap saved as a pdf file.
}
\description{
This function takes in a data matrix and produces an annotated heatmap using the pheatmap package.
}
\author{
Daniel Braas
}
