% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medium.module.R
\name{medium_analysis}
\alias{medium_analysis}
\title{This is the corresponding server function to medium_analysisUI.}
\usage{
medium_analysis(input, output, session, volumes = volumes)
}
\value{
The processed data in the selected folder.
}
\description{
This is the corresponding server function to medium_analysisUI.
}
\author{
Daniel Braas
}
