% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medium.module.R
\name{medium_analysisUI}
\alias{medium_analysisUI}
\title{This function produces the UI to process medium data analyzed with TraceFinder.}
\usage{
medium_analysisUI(id)
}
\arguments{
\item{id}{The module id used in the app.}
}
\value{
The processed data in the selected folder.
}
\description{
This function produces the UI to process medium data analyzed with TraceFinder.
}
\author{
Daniel Braas
}
