% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{medium_relative}
\alias{medium_relative}
\title{medium_relative subtracts the blank/fresh medium samples from all samples which contain
the corresponding medium. The output is a data frame and bar graphs of
all metabolites where negative numbers represent consumed metabolites
and positive numbers produced metabolites.}
\usage{
medium_relative(dat, Title = "", info)
}
\arguments{
\item{dat}{A formated data frame produced by prepare_data().}

\item{Title}{Project title.}

\item{info}{A data frame with information about the experiment.}
}
\value{
A data frame with integrated metabolomics data as well as
bar graphs of all measured metabolites.
}
\description{
medium_relative subtracts the blank/fresh medium samples from all samples which contain
the corresponding medium. The output is a data frame and bar graphs of
all metabolites where negative numbers represent consumed metabolites
and positive numbers produced metabolites.
}
\author{
Daniel Braas
}
