% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{prepare_data}
\alias{prepare_data}
\title{prepare_data is a function gathers the individual integration files produced by TraceFinder.}
\usage{
prepare_data(
  data,
  Title,
  Sample.Name = "",
  numbers = 1,
  std = FALSE,
  bkg = FALSE,
  norm = "none"
)
}
\arguments{
\item{data}{A formated data frame that contains the non-normalized MS data.}

\item{Title}{Project title}

\item{Sample.Name}{A vector containing the names for each sample. Usually,
this will be a concatenation of the condition and the sample repeat.}

\item{numbers}{A vector of cell numbers or other specimen normalization
quantity.}

\item{std}{(boolean) Should a plot of all internal standards be created.}

\item{bkg}{(boolean) Should the background be subtracted from all samples.}

\item{norm}{The standard (internal or external) to normalize to.}
}
\value{
A data frame with integrated metabolomics data.
}
\description{
prepare_data is a function gathers the individual integration files produced by TraceFinder.
}
\author{
Daniel Braas
}
