% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{shape_raw}
\alias{shape_raw}
\title{This function separates the compound names into compound and isotopologue names.}
\usage{
shape_raw(dat)
}
\arguments{
\item{dat}{A data frame with integrated metabolomics data.}
}
\value{
A data frame where the compound names and isotopologue data have been separated and formated.
}
\description{
This function separates the compound names into compound and isotopologue names.
}
\author{
Daniel Braas
}
